+++ smshub/lib/smshub.lib.php
@@ -15,10 +15,6 @@ if (!function_exists('smshubAdminTabs')) {
 		$head[$h][1] = $langs->trans("SmsHubTemplates");
 		$head[$h][2] = 'templates';
 		$h++;
-		$head[$h][0] = dol_buildpath('/smshub/admin/relances.php', 1);
-		$head[$h][1] = $langs->trans("SmsHubRelances");
-		$head[$h][2] = 'relances';
-		$h++;
 		$head[$h][0] = dol_buildpath('/smshub/admin/log.php', 1);
 		$head[$h][1] = $langs->trans("SmsHubLog");
 		$head[$h][2] = 'log';