+++ smshub/sql/llx_smshub_relance_status.key.sql
@@ -1,1 +1,0 @@
-ALTER TABLE llx_smshub_relance_status ADD UNIQUE INDEX uk_smshub_relance_status_fk (fk_facture, entity);