+++ smshub/sql/llx_smshub_relance_step.key.sql
@@ -1,1 +1,0 @@
-ALTER TABLE llx_smshub_relance_step ADD UNIQUE INDEX uk_smshub_relance_step (rank_order, entity);